脚本上传

+++ mogu/uploadscript.py
@@ -3,6 +3,7 @@
 #Time: 下午10:28
 from mogu.models.model import Plugin, PluginVersion, Images, WebSiteUrl, Kind
 from google.appengine.ext.blobstore import blobstore, BlobInfo
+from google.appengine.api import files
 from google.appengine.ext import db
 from tools.page import Page
 import datetime
@@ -23,11 +24,15 @@ class PluginUploadScript(Page):
                     for img in Images.get_by_id(pv.imageids):
                         if img:
                             img.delete()
-                    blobstore.BlobInfo.get(pv.datakey).delete()
+                    if pv.datakey:
+                        b=blobstore.BlobInfo.get(pv.datakey)
+                        if b:
+                            b.delete()
                     pv.delete()
-                img = Images.get_by_id(plugin.imageid)
-                if img:
-                    img.delete()
+                if plugin.imageid:
+                    img = Images.get_by_id(plugin.imageid)
+                    if img:
+                        img.delete()
                 plugin.delete()
         desc = self.request.get('desc', None)
         if id:
@@ -84,5 +89,39 @@ class PluginUploadScript(Page):
                 pluginVersion.imageids.append(imgfile.key().id())
         pluginVersion.put()
 
-        upload_url = blobstore.create_upload_url('%s/upload?pluginversionid=%s'%(WebSiteUrl.getInitUrl(),pluginVersion.key().id()))
+        upload_url = blobstore.create_upload_url('/upload?pluginversionid=%s'%(pluginVersion.key().id()))
         self.flush({'status' : 200, 'upload_url' : upload_url})
+
+
+
+class PluginUploadApkScript(Page):
+    def post(self):
+        appcode = self.request.get('appcode', None)
+        for plugin in Plugin.all().filter('appcode =', appcode):
+            for pv in PluginVersion.all().filter('plugin =',plugin.key().id()):
+                if pv.datakey:
+                    b=blobstore.BlobInfo.get(pv.datakey)
+                    if b:
+                        b.delete()
+                upload_url = blobstore.create_upload_url('/upload?pluginversionid=%s'%(pv.key().id()))
+                return self.flush({'status' : 200, 'upload_url' : upload_url})
+        return self.flush({'status':500,'message':u'插件已经存在'})
+
+
+class PluginUploadApkDataScript(Page):
+    def post(self):
+        pluginversionid = self.request.get('pluginversionid', None)
+        pv = PluginVersion.get_by_id(int(pluginversionid))
+        if pv.datakey:
+            b = BlobInfo.get(pv.datakey)
+            if b:
+                b.delete()
+        file_name = files.blobstore.create(mime_type='application/octet-stream')
+        with files.open(file_name, 'a') as f:
+            f.write(self.request.POST.get('file'))
+        files.finalize(file_name)
+        blob_key = files.blobstore.get_blob_key(file_name)
+        pv.datakey=str(blob_key)
+        pv.size = files.blobstore.size
+        pv.save()
+        return self.flush({'status':200,'message':u'插件已经存在'})
